// crawler/src/database/hasuraAPI/graphql/generated.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  executionstatus: any;
  time: any;
};
// ...
export type InsertTagMutationVariables = Exact<{
  tag_name?: InputMaybe<Scalars['String']>;
}>;


export type InsertTagMutation = { __typename?: 'mutation_root', insert_tags_one?: { __typename?: 'tags', id: number } | null };

export type AllTagsQueryVariables = Exact<{ [key: string]: never; }>;


export type AllTagsQuery = { __typename?: 'query_root', tags: Array<{ __typename?: 'tags', id: number, tag_name: string }> };

export type AllTagsInListQueryVariables = Exact<{
  _in?: InputMaybe<Array<Scalars['Int']> | Scalars['Int']>;
}>;


export type AllTagsInListQuery = { __typename?: 'query_root', tags: Array<{ __typename?: 'tags', id: number }> };

export type CountOfTagsInListQueryVariables = Exact<{
  _in?: InputMaybe<Array<Scalars['Int']> | Scalars['Int']>;
}>;


export type CountOfTagsInListQuery = { __typename?: 'query_root', tags_aggregate: { __typename?: 'tags_aggregate', aggregate?: { __typename?: 'tags_aggregate_fields', count: number } | null } };

export type TagsRecordRelationsByRecordIdQueryVariables = Exact<{
  recordId?: InputMaybe<Scalars['Int']>;
}>;


export type TagsRecordRelationsByRecordIdQuery = { __typename?: 'query_root', tags_records_relations: Array<{ __typename?: 'tags_records_relations', record_id: number, id: number, tag_id: number }> };
// ...
export type GetNumberOfTagsQuery = { __typename?: 'query_root', tags_aggregate: { __typename?: 'tags_aggregate', aggregate?: { __typename?: 'tags_aggregate_fields', count: number } | null } };